Video 11 complete, Apply Permissions

+++ backend/resources/views/articles/list.blade.php
@@ -4,7 +4,9 @@
         <h2 class="font-semibold text-xl text-gray-800 dark:text-gray-200 leading-tight">
             {{ __('Articles') }}
         </h2>
-        <a href="{{route('articles.create')}}" class="bg-slate-700 text-sm rounded-md text-white px-3 py-3">Create</a>
+        @can('create articles')
+            <a href="{{route('articles.create')}}" class="bg-slate-700 text-sm rounded-md text-white px-3 py-3">Create</a>
+        @endcan
         </div>
     </x-slot>
 
@@ -40,8 +42,14 @@
                             {{\Carbon\Carbon::parse($article->created_at)->format('d M, Y')}}
                         </td>
                         <td class="px-6 py-3 text-center">
-                            <a href="{{ route('articles.edit', $article->id) }}" class="bg-slate-700 text-sm rounded-md text-white px-3 py-2 hover:bg-slate-600">Edit</a>
-                            <a href="javascript:void(0);" onclick="deleteArticle({{ $article->id }})" class="bg-red-700 text-sm rounded-md text-white px-3 py-2 hover:bg-red-600">Delete</a>
+
+                            @can('edit articles')
+                                <a href="{{ route('articles.edit', $article->id) }}" class="bg-slate-700 text-sm rounded-md text-white px-3 py-2 hover:bg-slate-600">Edit</a>
+                            @endcan
+                            @can('delete articles')
+                                <a href="javascript:void(0);" onclick="deleteArticle({{ $article->id }})" class="bg-red-700 text-sm rounded-md text-white px-3 py-2 hover:bg-red-600">Delete</a>
+                            @endcan
+
                         </td>
 
                     </tr>
