+++ backend/resources/views/users/list.blade.php
@@ -46,8 +46,12 @@
                             {{\Carbon\Carbon::parse($user->created_at)->format('d M, Y')}}
                         </td>
                         <td class="px-6 py-3 text-center">
-                        <a href="{{ route('users.edit', $user->id) }}" class="bg-slate-700 text-sm rounded-md text-white px-3 py-2 hover:bg-slate-600">Edit</a>
-                        <a href="javascript:void(0);" onclick="deleteRole({{ $user->id }})" class="bg-red-700 text-sm rounded-md text-white px-3 py-2 hover:bg-red-600">Delete</a>
+                        @can('edit users')
+                            <a href="{{ route('users.edit', $user->id) }}" class="bg-slate-700 text-sm rounded-md text-white px-3 py-2 hover:bg-slate-600">Edit</a>
+                        @endcan
+                        @can('delete users')
+                            <a href="javascript:void(0);" onclick="deleteRole({{ $user->id }})" class="bg-red-700 text-sm rounded-md text-white px-3 py-2 hover:bg-red-600">Delete</a>
+                        @endcan
                         </td>
 
                     </tr>
